pyproject.toml
[project]
name = "gemini-live-avatar"
version = "0.1.6"
description = "Create a live avatar using Gemini and FastAPI"
readme = "README.md"
authors = [
    { name = "haruiz", email = "henryruiz22@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.16.1",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "typer>=0.16.0",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pdm.scripts]
gemini-live-avatar = "gemini_live_avatar.cli:app"

[project.scripts]
gemini-live-avatar = "gemini_live_avatar.cli:app"

[project.optional-dependencies]
mcp-demo = [
    "fastapi-mcp>=0.3.4",
    "mcp[cli]>=1.9.2",
]

[tool.hatch.build]
exclude = [
    "tests/",
    "docs/",
    "examples/",
    "assets/",
    "keys/",
    ".git/",
    ".github/",
    "scripts/",
    "configs/",
    "README.md",
    ".env",
    "config.json",
    "*/**/.sh",
    "Makefile",
    "cloudbuild.yaml",
    "*/**/Dockerfile",
    "Dockerfile",
    "requirements.txt",
    "docker/",
    "*.sh",
    "*.yml",
    "analysis/",
    "data/",
    "images/",
    "ui/",
    "web.ps1",
    ".vscode/",
    ".idea/",
    ".gitignore",
    ".gitattributes",
    ".gitmodules",
    ".gitkeep",
    ".gitlab-ci.yml",
    ".gitlab/",
    ".gitlab-ci/",
    ".gitlab-ci.yml",
    ".python-version",
    "uv.lock"

]
